requirements.txt
meshtastic
paho-mqtt
matrix-nio
python-dotenv
asyncio
cryptography
